# backend/docs/swagger.yaml
definitions:
...
  domain.ChangeUserAccess:
    properties:
      id:
        type: integer
      new_access:
        type: integer
    required:
    - id
    - new_access
    type: object
  domain.ComplaintID:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  domain.LevelBan:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  domain.LevelComplaint:
    properties:
      author_id:
        type: integer
      level_id:
        type: integer
      message:
        type: string
      reason:
        type: string
    required:
    - author_id
    - level_id
    - message
    - reason
    type: object
  domain.LevelComplaints:
    properties:
      level_complaints:
        items:
          $ref: '#/definitions/domain.LevelComplaint'
        type: array
    type: object
...
  domain.User:
    properties:
      name:
        type: string
      password:
        description: hash
        type: string
    required:
    - name
    - password
    type: object
  domain.UserBan:
    properties:
      ban_reason:
        type: string
      ban_time:
        type: string
      id:
        type: integer
    required:
    - ban_reason
    - ban_time
    - id
    type: object
  domain.UserComplaint:
    properties:
      author_id:
        type: integer
      message:
        type: string
      reason:
        type: string
      user_id:
        type: integer
    required:
    - author_id
    - message
    - reason
    - user_id
    type: object
  domain.UserComplaints:
    properties:
      user_complaints:
        items:
          $ref: '#/definitions/domain.UserComplaint'
        type: array
    type: object
...
  domain.UserSearchParams:
    properties:
      is_banned:
        type: boolean
      name:
        type: string
      offset:
        type: integer
      page_size:
        type: integer
    required:
    - is_banned
    - name
    - offset
    - page_size
    type: object
  domain.UserUnban:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  domain.Users:
    properties:
      users:
        items:
          $ref: '#/definitions/domain.User'
        type: array
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for GoType game and website
  title: GoType App API
  version: 0.0.1
paths:
  /admin/ban-level:
    post:
      consumes:
      - application/json
      description: Ban level with given id. Available only for moderators and admins
      operationId: ban-level
      parameters:
      - description: id of level you want to ban
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.LevelBan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_ENTITY_NOT_FOUND - Level with such
            id does not exist; ERR_INVALID_INPUT - Wrong structure of input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Ban level
      tags:
      - admin
  /admin/ban-user:
    post:
      consumes:
      - application/json
      description: Ban user with given id if your access is greater than theirs. Available
        only for moderators and admins
      operationId: ban-user
      parameters:
      - description: id of user you want to ban, duration of ban (format 10h), ban_reason
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UserBan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_NO_SUCH_USER - User with such id
            does not exist; ERR_INVALID_INPUT - Wrong structure of input json/Wrong
            format of ban duration;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Ban user
      tags:
      - admin
  /admin/change-user-access:
    post:
      consumes:
      - application/json
      description: Change access of a user if your access is greater that theirs.
        New value must be less than admin's one. Available only for moderators and
        admins
      operationId: change-user-access
      parameters:
      - description: id of user you want to ban, new value of access
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.ChangeUserAccess'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_NO_SUCH_USER - User with such id
            does not exist; ERR_INVALID_INPUT - Wrong structure of input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Change access level
      tags:
      - admin
  /admin/get-level-complaints:
    get:
      consumes:
      - application/json
      description: Get list of level complaints assigned to current admin. Available
        only for moderators and admins
      operationId: get-level-complaints
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LevelComplaints'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Level Complaints for moderator to process
      tags:
      - admin
  /admin/get-user-complaints:
    get:
      consumes:
      - application/json
      description: Get list of user complaints assigned to current admin. Available
        only for moderators and admins
      operationId: get-user-complaints
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserComplaints'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get User Complaints for moderator to process
      tags:
      - admin
  /admin/get-users:
    get:
      consumes:
      - application/json
      description: Get list of users with given params. Available only for moderators
        and admins
      operationId: get-users
      parameters:
      - description: Search params - name, isBanned, pages info
        in: body
        name: user_search_params
        required: true
        schema:
          $ref: '#/definitions/domain.UserSearchParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Users'
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Users with given params
      tags:
      - admin
  /admin/process-level-complaint:
    post:
      consumes:
      - application/json
      description: Delete level complaint with given id. Available only for moderators
        and admins
      operationId: process-level-complaint
      parameters:
      - description: Complaint ID
        in: body
        name: complaint_id
        required: true
        schema:
          $ref: '#/definitions/domain.ComplaintID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_ENTITY_NOT_FOUND - There is no complaint
            with such id among the ones assigned to you'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete level complaint with given id
      tags:
      - admin
  /admin/process-user-complaint:
    post:
      consumes:
      - application/json
      description: Delete user complaint with given id. Available only for moderators
        and admins
      operationId: process-user-complaint
      parameters:
      - description: Complaint ID
        in: body
        name: complaint_id
        required: true
        schema:
          $ref: '#/definitions/domain.ComplaintID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_ENTITY_NOT_FOUND - There is no complaint
            with such id among the ones assigned to you'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete user complaint with given id
      tags:
      - admin
  /admin/unban-level:
    post:
      consumes:
      - application/json
      description: Unban level with given id. Available only for moderators and admins
      operationId: unban-level
      parameters:
      - description: id of level you want to unban
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.LevelBan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_ENTITY_NOT_FOUND - Level with such
            id does not exist; ERR_INVALID_INPUT - Wrong structure of input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Unban level
      tags:
      - admin
  /admin/unban-user:
    post:
      consumes:
      - application/json
      description: Unban user with given id if your access is greater than theirs.
        Available only for moderators and admins
      operationId: unban-user
      parameters:
      - description: id of user you want to unban
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UserUnban'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: 'Possible messages: ERR_ACCESS_TOKEN_WRONG - Wrong structure
            of Access Token/No Access Token; ERR_NO_SUCH_USER - User with such id
            does not exist; ERR_INVALID_INPUT - Wrong structure of input json;'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: 'Possible messages: ERR_UNAUTHORIZED - Access Token expired;
            ERR_PERMISSION_DENIED - Not enough rights to perform the action'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: 'Possible messages: ERR_INTERNAL - Error on server'
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Unban user
      tags:
      - admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"